Refactoring and grouping arithmetic expressions
