+++ smith/src/program/expr/expr.rs
@@ -1,14 +1,11 @@
-use crate::program::types::TypeID;
 use crate::program::var::Var;
 
+use super::arithmetic_expr::{ArithmeticExpr, IntExpr};
+use super::bool_expr::BoolExpr;
 use super::borrow_expr::BorrowExpr;
-use super::func_call_expr::FunctionCallExpr;
-use super::int_expr::IntExpr;
 use super::struct_expr::StructExpr;
-use super::{binary_expr::BinaryExpr, bool_expr::BoolExpr};
 
-use strum_macros::{EnumCount, EnumDiscriminants, EnumIter};
-
+// The top most form of an expression
 pub enum Expr {
     Arithmetic(ArithmeticExpr),
     Boolean(BoolExpr),
@@ -29,6 +26,7 @@ impl ToString for Expr {
     }
 }
 
+// Literally
 pub enum LiteralExpr {
     Int(IntExpr),
     Struct(StructExpr),
@@ -44,51 +42,9 @@ impl ToString for LiteralExpr {
         }
     }
 }
-#[derive(EnumDiscriminants)]
-#[strum_discriminants(vis(pub))]
-#[strum_discriminants(name(ArithmeticExprVariants))]
-#[strum_discriminants(derive(EnumCount, EnumIter))]
-pub enum ArithmeticExpr {
-    Int(IntExpr),
-    Binary(Box<BinaryExpr>),
-    Var(Var),
-    Func(FunctionCallExpr),
-}
 
-impl ArithmeticExpr {
-    pub fn new_from_bin_expr(expr: BinaryExpr) -> Self {
-        ArithmeticExpr::Binary(Box::new(expr))
-    }
-
-    pub fn new_from_int_expr(expr: IntExpr) -> Self {
-        ArithmeticExpr::Int(expr)
-    }
-
-    pub fn as_expr(self) -> Expr {
-        Expr::Arithmetic(self)
-    }
-
-    pub fn get_type(&self) -> TypeID {
-        match self {
-            Self::Int(s) => s.get_type(),
-            Self::Binary(s) => s.get_type(),
-            Self::Var(s) => s.get_type(),
-            Self::Func(s) => s.get_type(),
-        }
-    }
-}
-
-impl ToString for ArithmeticExpr {
-    fn to_string(&self) -> String {
-        match self {
-            Self::Int(s) => s.to_string(),
-            Self::Binary(s) => (*s).to_string_safe(),
-            Self::Var(s) => s.to_string(),
-            Self::Func(s) => s.to_string(),
-        }
-    }
-}
-
+// Using string only to represent the expression
+// This should not be used unless needed as it doesn't reflect the program's AST
 pub struct RawExpr {
     expr_string: String,
 }
