+++ smith/src/program/expr/mod.rs
@@ -1,7 +1,6 @@
-pub mod binary_expr;
+pub mod arithmetic_expr;
 pub mod bool_expr;
 pub mod borrow_expr;
 pub mod expr;
 pub mod func_call_expr;
-pub mod int_expr;
 pub mod struct_expr;
